Shrink the sizes of some constants
This commit changes several constants to use smaller types. This means
we don't have to downcast them to use them, and we can safely convert
them to larger types when needed.

One limitation with this is that Kani doesn't support casts in length
fields, so I changed those cases to use a separate constant containing
the cast.

+++ src/seccompiler/src/backend.rs
@@ -725,7 +725,7 @@ impl SeccompFilter {
         accumulator.push(built_syscall);
 
         // BPF programs are limited to 4096 statements.
-        if *filter_len >= BPF_MAX_LEN {
+        if *filter_len >= usize::from(BPF_MAX_LEN) {
             return Err(FilterError::FilterTooLarge);
         }
 
@@ -781,7 +781,7 @@ impl TryInto<BpfProgram> for SeccompFilter {
             .into_iter()
             .for_each(|mut instructions| result.append(&mut instructions));
 
-        if result.len() >= BPF_MAX_LEN {
+        if result.len() >= usize::from(BPF_MAX_LEN) {
             return Err(FilterError::FilterTooLarge);
         }
